rename lcapy.ts to python.ts
